backend/requirements.txt
anyio==4.12.1
PyJWT==2.10.1
APScheduler==3.11.2
beautifulsoup4==4.14.3
fastapi==0.135.1
httpx==0.28.1
lxml==5.3.2
pandas==3.0.1
pandas-market-calendars==5.3.1.2
pandas-ta==0.4.71b0
pydantic==2.12.5
python-dotenv==1.2.2
requests==2.32.5
supabase==2.28.3
uvicorn==0.42.0
yfinance==1.2.0
